+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Ignore discord server errors 503 and alert on too many of them.
+
 ## [v18.4.10](https://github.com/lexicalunit/spellbot/releases/tag/v18.4.10) - 2026-04-22
 
 ### Changed